feature(imports) support jwt config for get all reasource in meta
