+++ config.go
@@ -35,6 +35,7 @@ type Config struct {
 	Environment              string           `mapstructure:"ENVIRONMENT"`
 	GoogleProjectId          string           `mapstructure:"GOOGLE_PROJECT_ID"`
 	GoogleSaPath             string           `mapstructure:"GOOGLE_SA_PATH"`
+	JwtToken                 string           `mapstructure:"JWT_TOKEN"`
 	MaxServerRequestBodySize int              `mapstructure:"MAX_SERVER_REQUEST_BODY_SIZE"`
 	Mode                     Mode             `mapstructure:"MODE"`
 	PgMetaUrl                string           `mapstructure:"PG_META_URL"`
