+++ pkg/connector/pgmeta/index.go
@@ -10,7 +10,7 @@ import (
 
 func GetIndexes(cfg *raiden.Config, schema string) ([]objects.Index, error) {
 	MetaLogger.Trace("start fetching indexes from meta")
-	rs, err := ExecuteQuery[[]objects.Index](cfg.PgMetaUrl, sql.GenerateGetIndexQuery(schema), nil, nil, nil)
+	rs, err := ExecuteQuery[[]objects.Index](cfg.PgMetaUrl, sql.GenerateGetIndexQuery(schema), nil, DefaultAuthInterceptor(cfg.JwtToken), nil)
 	if err != nil {
 		err = fmt.Errorf("get indexes error : %s", err)
 		return []objects.Index{}, err
