+++ pkg/connector/pgmeta/table.go
@@ -1,7 +1,6 @@
 package pgmeta
 
 import (
-	"encoding/json"
 	"errors"
 	"fmt"
 	"net/http"
@@ -11,30 +10,10 @@ import (
 
 	"github.com/sev-2/raiden"
 	"github.com/sev-2/raiden/pkg/client/net"
-	"github.com/sev-2/raiden/pkg/logger"
 	"github.com/sev-2/raiden/pkg/supabase/objects"
 	"github.com/sev-2/raiden/pkg/supabase/query"
 	"github.com/sev-2/raiden/pkg/supabase/query/sql"
 )
-
-var MetaLogger = logger.HcLog().Named("connector.meta")
-
-// ----- Execute Query -----
-type ExecuteQueryParam struct {
-	Query     string `json:"query"`
-	Variables any    `json:"variables"`
-}
-
-func ExecuteQuery[T any](baseUrl, query string, variables any, reqInterceptor net.RequestInterceptor, resInterceptor net.ResponseInterceptor) (result T, err error) {
-	url := fmt.Sprintf("%s/query", baseUrl)
-	p := ExecuteQueryParam{Query: query, Variables: variables}
-	pByte, err := json.Marshal(p)
-	if err != nil {
-		MetaLogger.Error("error execute query", "query", query)
-		return result, err
-	}
-	return net.Post[T](url, pByte, net.DefaultTimeout, reqInterceptor, resInterceptor)
-}
 
 func GetTables(cfg *raiden.Config, includedSchemas []string, includeColumns bool) ([]objects.Table, error) {
 	MetaLogger.Trace("start fetching tables from meta")
@@ -47,6 +26,10 @@ func GetTables(cfg *raiden.Config, includedSchemas []string, includeColumns bool
 
 		if includeColumns {
 			req.URL.Query().Set("include_columns", strconv.FormatBool(includeColumns))
+		}
+
+		if len(cfg.JwtToken) > 0 {
+			req.Header.Set("Authorization", fmt.Sprintf("Bearer %s", cfg.JwtToken))
 		}
 
 		return nil
@@ -68,7 +51,7 @@ func GetTableByName(cfg *raiden.Config, name, schema string, includeColumn bool)
 		return result, err
 	}
 
-	rs, err := ExecuteQuery[[]objects.Table](cfg.PgMetaUrl, q, nil, nil, nil)
+	rs, err := ExecuteQuery[[]objects.Table](cfg.PgMetaUrl, q, nil, DefaultAuthInterceptor(cfg.JwtToken), nil)
 	if err != nil {
 		err = fmt.Errorf("get tables error : %s", err)
 		return
@@ -95,7 +78,7 @@ func CreateTable(cfg *raiden.Config, newTable objects.Table) (result objects.Tab
 	}
 
 	// execute update
-	_, err = ExecuteQuery[any](cfg.PgMetaUrl, sql, nil, nil, nil)
+	_, err = ExecuteQuery[any](cfg.PgMetaUrl, sql, nil, DefaultAuthInterceptor(cfg.JwtToken), nil)
 	if err != nil {
 		return result, fmt.Errorf("create new table %s error : %s", newTable.Name, err)
 	}
@@ -107,7 +90,7 @@ func UpdateTable(cfg *raiden.Config, newTable objects.Table, updateItem objects.
 	MetaLogger.Trace("start update table", "name", newTable.Name)
 	sql := query.BuildUpdateTableQuery(newTable, updateItem)
 	// execute update
-	_, err := ExecuteQuery[any](cfg.PgMetaUrl, sql, nil, nil, nil)
+	_, err := ExecuteQuery[any](cfg.PgMetaUrl, sql, nil, DefaultAuthInterceptor(cfg.JwtToken), nil)
 	if err != nil {
 		return fmt.Errorf("update tables error : %s", err)
 	}
@@ -142,7 +125,7 @@ func DeleteTable(cfg *raiden.Config, table objects.Table, cascade bool) error {
 	MetaLogger.Trace("start delete table", "name", table.Name)
 	sql := query.BuildDeleteTableQuery(table, true)
 	// execute delete
-	_, err := ExecuteQuery[any](cfg.PgMetaUrl, sql, nil, nil, nil)
+	_, err := ExecuteQuery[any](cfg.PgMetaUrl, sql, nil, DefaultAuthInterceptor(cfg.JwtToken), nil)
 	if err != nil {
 		return fmt.Errorf("delete table %s error : %s", table.Name, err)
 	}
@@ -155,7 +138,7 @@ func GetTableRelationshipActions(cfg *raiden.Config, schema string) ([]objects.T
 	MetaLogger.Trace("start fetching table relationships from pg-meta")
 	q := sql.GenerateGetTableRelationshipActionsQuery(schema)
 
-	rs, err := ExecuteQuery[[]objects.TablesRelationshipAction](cfg.PgMetaUrl, q, nil, nil, nil)
+	rs, err := ExecuteQuery[[]objects.TablesRelationshipAction](cfg.PgMetaUrl, q, nil, DefaultAuthInterceptor(cfg.JwtToken), nil)
 	if err != nil {
 		return rs, fmt.Errorf("get tables relation from pg-meta error : %s", err)
 	}
@@ -265,7 +248,7 @@ func CreateColumn(cfg *raiden.Config, column objects.Column, isPrimary bool) err
 	}
 
 	// Execute SQL Query
-	_, err = ExecuteQuery[any](cfg.PgMetaUrl, sql, nil, nil, nil)
+	_, err = ExecuteQuery[any](cfg.PgMetaUrl, sql, nil, DefaultAuthInterceptor(cfg.JwtToken), nil)
 	if err != nil {
 		return fmt.Errorf("create column %s.%s error : %s", column.Table, column.Name, err)
 	}
@@ -279,7 +262,7 @@ func UpdateColumn(cfg *raiden.Config, oldColumn, newColumn objects.Column, updat
 	sql := query.BuildUpdateColumnQuery(oldColumn, newColumn, updateItem)
 
 	// Execute SQL Query
-	_, err := ExecuteQuery[any](cfg.PgMetaUrl, sql, nil, nil, nil)
+	_, err := ExecuteQuery[any](cfg.PgMetaUrl, sql, nil, DefaultAuthInterceptor(cfg.JwtToken), nil)
 	if err != nil {
 		return fmt.Errorf("update column %s.%s error : %s", newColumn.Table, newColumn.Name, err)
 	}
@@ -290,7 +273,7 @@ func UpdateColumn(cfg *raiden.Config, oldColumn, newColumn objects.Column, updat
 func DeleteColumn(cfg *raiden.Config, column objects.Column) error {
 	MetaLogger.Trace("start delete column", "table", column.Table, "name", column.Name)
 	sql := query.BuildDeleteColumnQuery(column)
-	_, err := ExecuteQuery[any](cfg.PgMetaUrl, sql, nil, nil, nil)
+	_, err := ExecuteQuery[any](cfg.PgMetaUrl, sql, nil, DefaultAuthInterceptor(cfg.JwtToken), nil)
 	if err != nil {
 		return fmt.Errorf("delete column %s.%s error : %s", column.Table, column.Name, err)
 	}
@@ -397,7 +380,7 @@ func createForeignKey(cfg *raiden.Config, relation *objects.TablesRelationship) 
 		return err
 	}
 
-	_, err = ExecuteQuery[any](cfg.PgMetaUrl, sql, nil, nil, nil)
+	_, err = ExecuteQuery[any](cfg.PgMetaUrl, sql, nil, DefaultAuthInterceptor(cfg.JwtToken), nil)
 	if err != nil {
 		return fmt.Errorf("create foreign key %s.%s error : %s", relation.SourceTableName, relation.SourceColumnName, err)
 	}
@@ -406,7 +389,7 @@ func createForeignKey(cfg *raiden.Config, relation *objects.TablesRelationship) 
 	if indexSql, err := query.BuildFKIndexQuery(objects.UpdateRelationCreate, relation); err != nil {
 		return err
 	} else if len(indexSql) > 0 {
-		_, err = ExecuteQuery[any](cfg.PgMetaUrl, indexSql, nil, nil, nil)
+		_, err = ExecuteQuery[any](cfg.PgMetaUrl, indexSql, nil, DefaultAuthInterceptor(cfg.JwtToken), nil)
 		if err != nil {
 			return fmt.Errorf("create foreign index %s.%s error : %s", relation.SourceTableName, relation.SourceColumnName, err)
 		}
@@ -429,7 +412,7 @@ func updateForeignKey(cfg *raiden.Config, relation *objects.TablesRelationship) 
 	}
 
 	sql := deleteSql + createSql
-	_, err = ExecuteQuery[any](cfg.PgMetaUrl, sql, nil, nil, nil)
+	_, err = ExecuteQuery[any](cfg.PgMetaUrl, sql, nil, DefaultAuthInterceptor(cfg.JwtToken), nil)
 	if err != nil {
 		return fmt.Errorf("update foreign key %s.%s error : %s", relation.SourceTableName, relation.SourceColumnName, err)
 	}
@@ -439,7 +422,7 @@ func updateForeignKey(cfg *raiden.Config, relation *objects.TablesRelationship) 
 		if indexSql, err := query.BuildFKIndexQuery(objects.UpdateRelationCreate, relation); err != nil {
 			return err
 		} else if len(indexSql) > 0 {
-			_, err = ExecuteQuery[any](cfg.PgMetaUrl, indexSql, nil, nil, nil)
+			_, err = ExecuteQuery[any](cfg.PgMetaUrl, indexSql, nil, DefaultAuthInterceptor(cfg.JwtToken), nil)
 			if err != nil {
 				return fmt.Errorf("create foreign index %s.%s error : %s", relation.SourceTableName, relation.SourceColumnName, err)
 			}
@@ -458,7 +441,7 @@ func deleteForeignKey(cfg *raiden.Config, relation *objects.TablesRelationship) 
 		return err
 	}
 
-	_, err = ExecuteQuery[any](cfg.PgMetaUrl, sql, nil, nil, nil)
+	_, err = ExecuteQuery[any](cfg.PgMetaUrl, sql, nil, DefaultAuthInterceptor(cfg.JwtToken), nil)
 	if err != nil {
 		return fmt.Errorf("delete foreign key %s.%s error : %s", relation.SourceTableName, relation.SourceColumnName, err)
 	}
@@ -468,7 +451,7 @@ func deleteForeignKey(cfg *raiden.Config, relation *objects.TablesRelationship) 
 		if indexSql, err := query.BuildFKIndexQuery(objects.UpdateRelationDelete, relation); err != nil {
 			return err
 		} else if len(indexSql) > 0 {
-			_, err = ExecuteQuery[any](cfg.PgMetaUrl, indexSql, nil, nil, nil)
+			_, err = ExecuteQuery[any](cfg.PgMetaUrl, indexSql, nil, DefaultAuthInterceptor(cfg.JwtToken), nil)
 			if err != nil {
 				return fmt.Errorf("delete foreign index %s.%s error : %s", relation.SourceTableName, relation.SourceColumnName, err)
 			}
